# feat: make aggkit-prover startup faster by 2x