--- a/crates/aggchain-proof-builder/Cargo.toml
+++ b/crates/aggchain-proof-builder/Cargo.toml
@@ -17,6 +17,7 @@ serde = { workspace = true, features = ["derive"] }
 sp1-core-executor.workspace = true
 sp1-prover.workspace = true
 sp1-sdk.workspace = true
+tracing.workspace = true
 
 aggchain-proof-contracts.workspace = true
 aggchain-proof-core.workspace = true