--- a/crates/aggkit-prover/src/main.rs
+++ b/crates/aggkit-prover/src/main.rs
@@ -1,6 +1,7 @@
 use aggkit_prover::version;
 use anyhow::Context as _;
 use clap::Parser as _;
+use prover_config::{CpuProverConfig, ProverType};
 use sp1_sdk::HashableKey as _;
 use sp1_zkvm::lib::utils::words_to_bytes_le;
 
@@ -31,8 +32,12 @@ fn main() -> anyhow::Result<()> {
             }
         }
         aggkit_prover::cli::Commands::Vkey => {
-            let vkey =
-                prover_executor::Executor::get_vkey(aggchain_proof_service::AGGCHAIN_PROOF_ELF);
+            let vkey = prover_executor::Executor::new(
+                &ProverType::CpuProver(CpuProverConfig::default()),
+                &None,
+                aggchain_proof_service::AGGCHAIN_PROOF_ELF,
+            )
+            .get_vkey();
             let vkey_hex = hex::encode(words_to_bytes_le(&vkey.hash_u32()));
 
             println!("aggchain_proof_vkey: 0x{vkey_hex}");