--- a/crates/agglayer-prover/src/lib.rs
+++ b/crates/agglayer-prover/src/lib.rs
@@ -1,7 +1,6 @@
 use std::{path::PathBuf, sync::Arc};
 
 use prover_engine::ProverEngine;
-use sp1_sdk::HashableKey;
 
 #[cfg(feature = "testutils")]
 pub mod fake;
@@ -50,11 +49,6 @@ pub fn main(cfg: PathBuf, version: &str, program: &'static [u8]) -> anyhow::Resu
         .start();
 
     Ok(())
-}
-
-pub fn get_vkey(program: &'static [u8]) -> String {
-    let vkey = prover_executor::Executor::get_vkey(program);
-    vkey.bytes32()
 }
 
 #[cfg(feature = "testutils")]