Startup goes from ~14s to ~7s on my machine.

The time is heavily dominated by building the SP1 provers: each prover build takes ~3s. This PR removes two prover builds, and only leaves two: one for the aggregation elf, and one for the aggchain proof elf.

We might be able to shave another 3s (so another 2x factor) by only having a single prover, but it would be harder to implement, and is thus left as future work for now.

Fixes #126

## PR Checklist:

- [x] I have performed a self-review of my code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] I have added or updated tests that comprehensively prove my change is effective or that my feature works